// src/shared/api/generated/schemas/auction.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Logistics Auctions API
 * OpenAPI spec version: 0.0.1
 */
import type { AuctionDates } from "./auctionDates";
import type { AuctionStatus } from "./auctionStatus";
import type { AuctionType } from "./auctionType";
import type { Cargo } from "./cargo";
import type { Organizer } from "./organizer";
import type { PaymentTerms } from "./paymentTerms";
import type { Point } from "./point";
import type { PrimaryAction } from "./primaryAction";
import type { Route } from "./route";
import type { Trading } from "./trading";
import type { UserTradingStatus } from "./userTradingStatus";

export interface Auction {
  uuid: string;
  requestNumber: string;
  type: AuctionType;
  status: AuctionStatus;
  userTradingStatus?: UserTradingStatus | null;
  hasMyBet?: boolean;
  primaryAction?: PrimaryAction | null;
  route: Route;
  dates: AuctionDates;
  cargo: Cargo;
  /** @nullable */
  currentPrice?: number | null;
  /** @nullable */
  availablePrice?: number | null;
  pricePerKm?: number;
  /** @nullable */
  betStep?: number | null;
  organizer: Organizer;
  points: Point[];
  paymentTerms: PaymentTerms;
  trading: Trading;
  isAvailable: boolean;
  isBidder: boolean;
}